# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoroBeacon API",
    "description": "Monitoring and alerting for Stellar/Soroban contract events. All endpoints are mounted under /api/v1. The API is read-write for monitors, rules and channels; alerts are read-only.",
    "version": "1.0.0",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/monitors": {
      "post": {
        "summary": "Create a monitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MonitorInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Monitor" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "get": {
        "summary": "List monitors",
        "parameters": [
          {
            "name": "enabled",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "When true, only enabled monitors are returned"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Opaque continuation cursor from a previous page"
          }
        ],
        "responses": {
          "200": {
            "description": "Monitors",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MonitorPage" }
              }
            }
          }
        }
      }
    },
    "/monitors/bulk": {
      "post": {
        "summary": "Bulk enable or disable monitors",
        "description": "Sets enabled on an explicit list of monitor IDs in one statement. Unknown IDs are returned and do not prevent the known ones from applying. There is no all-monitors shorthand.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BulkMonitorInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Applied",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BulkMonitorResult" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/monitors/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "get": {
        "summary": "Fetch one monitor",
        "responses": {
          "200": {
            "description": "Monitor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Monitor" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "summary": "Update a monitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MonitorInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Monitor" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Delete a monitor",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/monitors/{id}/rules": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "post": {
        "summary": "Add a rule to a monitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RuleInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Rule" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "get": {
        "summary": "List a monitor's rules",
        "responses": {
          "200": {
            "description": "Rules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Rule" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/monitors/{id}/rules/{ruleID}": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" },
        { "$ref": "#/components/parameters/RuleID" }
      ],
      "patch": {
        "summary": "Update a rule",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RuleInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Rule" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Delete a rule",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/channels": {
      "post": {
        "summary": "Create a notification channel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "get": {
        "summary": "List channels",
        "parameters": [
          {
            "name": "enabled",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "When true, only enabled channels are returned"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Opaque continuation cursor from a previous page"
          }
        ],
        "responses": {
          "200": {
            "description": "Channels",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelPage" }
              }
            }
          }
        }
      }
    },
    "/channels/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "get": {
        "summary": "Fetch one channel",
        "responses": {
          "200": {
            "description": "Channel",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "summary": "Update a channel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Delete a channel",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/channels/{id}/test": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "post": {
        "summary": "Send a test notification through a channel",
        "responses": {
          "200": { "description": "Test sent" },
          "404": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/alerts": {
      "get": {
        "summary": "List alerts",
        "parameters": [
          {
            "name": "monitor_id",
            "in": "query",
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "since",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "RFC 3339 lower bound on alert creation"
          },
          {
            "name": "until",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" },
            "description": "RFC 3339 upper bound on alert creation"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Opaque continuation cursor from a previous page"
          }
        ],
        "responses": {
          "200": {
            "description": "Alerts",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AlertPage" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" }
      },
      "RuleID": {
        "name": "ruleID",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" }
      }
    },
    "responses": {
      "Error": {
        "description": "Error envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable message"
          },
          "code": {
            "type": "string",
            "description": "Stable status token, e.g. Not Found"
          },
          "request_id": {
            "type": "string",
            "description": "Correlation ID, also on the X-Request-ID response header"
          },
          "details": {
            "type": "array",
            "description": "Field-level validation problems. Omitted when there are none so existing clients keep working. Each entry names a JSON path and why it failed; all problems in one request are listed together.",
            "items": { "$ref": "#/components/schemas/ErrorDetail" }
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": ["field", "reason"],
        "properties": {
          "field": {
            "type": "string",
            "description": "Dotted JSON path of the offending field, e.g. rules[0].params.min_amount"
          },
          "reason": {
            "type": "string",
            "description": "Human-readable failure for that field"
          }
        }
      },
      "Monitor": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "contract_ids": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Soroban contract strkeys (C...)"
          },
          "enabled": { "type": "boolean" },
          "channel_ids": {
            "type": "array",
            "items": { "type": "integer", "format": "int64" }
          }
        }
      },
      "MonitorInput": {
        "type": "object",
        "required": ["name", "contract_ids"],
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "contract_ids": {
            "type": "array",
            "minItems": 1,
            "items": { "type": "string", "pattern": "^C[A-Z2-7]{55}$" }
          },
          "enabled": { "type": "boolean", "default": true },
          "channel_ids": {
            "type": "array",
            "items": { "type": "integer", "format": "int64" }
          }
        }
      },
      "BulkMonitorInput": {
        "type": "object",
        "required": ["ids", "enabled"],
        "properties": {
          "ids": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": { "type": "integer", "format": "int64" }
          },
          "enabled": { "type": "boolean" }
        }
      },
      "BulkMonitorResult": {
        "type": "object",
        "required": ["updated", "unknown_ids", "enabled"],
        "properties": {
          "updated": { "type": "integer", "description": "How many existing monitors changed state" },
          "unknown_ids": {
            "type": "array",
            "items": { "type": "integer", "format": "int64" }
          },
          "enabled": { "type": "boolean" }
        }
      },
      "Rule": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "monitor_id": { "type": "integer", "format": "int64" },
          "type": {
            "type": "string",
            "enum": ["event_emitted", "value_threshold", "token_event"]
          },
          "params": { "type": "object" },
          "enabled": { "type": "boolean" }
        }
      },
      "RuleInput": {
        "type": "object",
        "required": ["type", "params"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["event_emitted", "value_threshold", "token_event"]
          },
          "params": {
            "type": "object",
            "description": "Type-specific; validated against the rule type"
          },
          "enabled": { "type": "boolean", "default": true }
        }
      },
      "Channel": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "type": {
            "type": "string",
            "enum": ["discord", "slack", "telegram", "email", "webhook"]
          },
          "config": { "type": "object" },
          "enabled": { "type": "boolean" }
        }
      },
      "ChannelInput": {
        "type": "object",
        "required": ["name", "type", "config"],
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "type": {
            "type": "string",
            "enum": ["discord", "slack", "telegram", "email", "webhook"]
          },
          "config": {
            "type": "object",
            "description": "Type-specific; the channel reference documents each"
          },
          "enabled": { "type": "boolean", "default": true }
        }
      },
      "MonitorPage": {
        "type": "object",
        "properties": {
          "monitors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Monitor" }
          },
          "next_cursor": {
            "type": "string",
            "description": "Absent on the last page"
          }
        }
      },
      "ChannelPage": {
        "type": "object",
        "properties": {
          "channels": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Channel" }
          },
          "next_cursor": {
            "type": "string",
            "description": "Absent on the last page"
          }
        }
      },
      "AlertPage": {
        "type": "object",
        "properties": {
          "alerts": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Alert" }
          },
          "next_cursor": {
            "type": "string",
            "description": "Absent on the last page"
          }
        }
      },
      "Alert": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "monitor_id": { "type": "integer", "format": "int64" },
          "monitor_name": { "type": "string" },
          "rule_id": { "type": "integer", "format": "int64" },
          "event_id": { "type": "string" },
          "event_name": { "type": "string" },
          "payload": { "type": "object" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}